pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "trackshot"
version = "0.0.1"
description = "Real-time ball detection and tracking on the edge using a custom-trained YOLOv8n model deployed on an ESP32-S3 microcontroller with a live web dashboard."
authors = [
  { name = "PierGit7" },
]
license = { file = "LICENSE" }
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "hydra-core>=1.3.2",
    "omegaconf>=2.3.0",
    "torch",
    "torchvision",
    "ultralytics",
    "pillow",
    "opencv-python-headless",
    "python-dotenv",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "ruff"
]

[tool.ruff]
line-length = 99
src = ["trackshot"]

[tool.ruff.lint]
extend-select = ["I"]

[tool.ruff.lint.isort]
known-first-party = ["trackshot"]
force-sort-within-sections = true
